Bump asahi-bless dependency to 0.4.2
This uses "/dev/mtd/by-name/nvram" as default nvram MTD partition
instead of the previous implicit default of "/dev/mtd0". This depends on
udev rules to create symlinks in "/dev/mtd/by-name/".
The ideal solution would probe for the correct MTD device. Since the
intention is to replace the direct writing to the nvram device with a
daemon handling nvram reads and writes spending effort on this is not
justified.